Added migrations reference, moved existing migrations to new namespace to not conflict with EF6 migrations
