+++ LobotJR.Test/LobotJR.Test.csproj
@@ -82,6 +82,7 @@
     <Compile Include="Systems\Fishing\TournamentSystemTests.cs" />
   </ItemGroup>
   <ItemGroup>
+    <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
