+++ LobotJR/LobotJR.csproj
@@ -82,6 +82,9 @@
     <Reference Include="System.Data.SQLite.EF6, Version=1.0.113.0, Culture=neutral, PublicKeyToken=db937bc2d44ff139, processorArchitecture=MSIL">
       <HintPath>..\packages\System.Data.SQLite.EF6.1.0.113.0\lib\net46\System.Data.SQLite.EF6.dll</HintPath>
     </Reference>
+    <Reference Include="System.Data.SQLite.EF6.Migrations, Version=1.0.112.0, Culture=neutral, processorArchitecture=MSIL">
+      <HintPath>..\packages\System.Data.SQLite.EF6.Migrations.1.0.112\lib\System.Data.SQLite.EF6.Migrations.dll</HintPath>
+    </Reference>
     <Reference Include="System.Data.SQLite.Linq, Version=1.0.113.0, Culture=neutral, PublicKeyToken=db937bc2d44ff139, processorArchitecture=MSIL">
       <HintPath>..\packages\System.Data.SQLite.Linq.1.0.113.0\lib\net46\System.Data.SQLite.Linq.dll</HintPath>
     </Reference>
@@ -118,8 +121,8 @@
     <Compile Include="IrcClient.cs" />
     <Compile Include="Items.cs" />
     <Compile Include="JSON.cs" />
-    <Compile Include="Migrations\FisherMigration.cs" />
-    <Compile Include="Migrations\FishMigration.cs" />
+    <Compile Include="Data\Import\FisherDataImport.cs" />
+    <Compile Include="Data\Import\FishDataImport.cs" />
     <Compile Include="Modules\AccessControl\AccessControlAdmin.cs" />
     <Compile Include="Modules\Fishing\FishingAdmin.cs" />
     <Compile Include="Modules\Fishing\FishingModule.cs" />
