+++ LobotJR/packages.config
@@ -7,5 +7,6 @@
   <package id="System.Data.SQLite" version="1.0.113.6" targetFramework="net462" />
   <package id="System.Data.SQLite.Core" version="1.0.113.6" targetFramework="net462" />
   <package id="System.Data.SQLite.EF6" version="1.0.113.0" targetFramework="net462" />
+  <package id="System.Data.SQLite.EF6.Migrations" version="1.0.112" targetFramework="net462" />
   <package id="System.Data.SQLite.Linq" version="1.0.113.0" targetFramework="net462" />
 </packages>
